# internal/server/invocation/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Windforce Invocation API",
    "version": "v1",
    "description": "Canonical system-to-system API for admitting and observing Runs. Authentication establishes the principal; callers cannot assert identity or execution environment fields."
  },
  "servers": [
    {
      "url": "https://windforce.example"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{workspace}/runs": {
      "post": {
        "operationId": "createRun",
        "summary": "Admit a Run",
        "parameters": [
          {
            "$ref": "#/components/parameters/Workspace"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An idempotent retry replayed the existing Run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "201": {
            "description": "Run admitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator, workspace, Client Registry (wfk_), or Service Principal (wfs_) token."
      }
    },
    "parameters": {
      "Workspace": {
        "name": "workspace",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Scopes retry-safe admission to the authenticated principal. Reusing a key with a different request returns 409.",
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      }
    },
    "schemas": {
      "CreateRunRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "app",
          "action",
          "input"
        ],
        "properties": {
          "app": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "input": true,
          "correlation_id": {
            "type": "string"
          }
        }
      },
...
      "Run": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "run_id",
          "state",
          "app",
          "action",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "waiting_human",
              "resuming",
              "succeeded",
              "failed",
              "canceled",
              "expired"
            ]
          },
          "app": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          },
          "replayed": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The principal lacks the required scope, ownership, or target permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Workspace, App, Action, or Run not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency or admission conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Invocation service unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}